+++ lib/src/api_client.dart
@@ -75,7 +75,7 @@ class ApiClient {
     final url = configuration.baseUrl + '/connect/token';
     final data = 'grant_type=client_credentials&client_id=${configuration.clientId}&client_secret=${configuration.clientSecret}';
     final headers = {'Content-Type' : 'application/x-www-form-urlencoded'};
-    final response = await http.post(url, headers: headers, body: data);
+    final response = await http.post(Uri.parse(url), headers: headers, body: data);
     if (response.statusCode == 400) {
       throw ApiException(response.statusCode, 'Invalid server credentials. Please check your ClientSecret and ClientId.');
     }
@@ -433,7 +433,7 @@ class ApiClient {
 
     var httpRequest = http.Request(requestData.method, Uri.parse(requestData.url));
     httpRequest.headers['x-aspose-client'] = 'dart sdk';
-    httpRequest.headers['x-aspose-client-version'] = '21.5';
+    httpRequest.headers['x-aspose-client-version'] = '21.6';
     httpRequest.headers['Authorization'] = await _getAuthToken();
     if (requestData.headers != null) {
       httpRequest.headers.addAll(requestData.headers);